# docs/api/openapi.yaml
components:
  schemas:
...
    api.SwaggerCommentResponse:
      properties:
        author_id:
          example: 874d6314-6353-45e9-ab2a-5fe930ea4dbc
          type: string
        author_name:
          example: Admin User
          type: string
        body:
          example: This looks good, let's merge it.
          type: string
        content:
          example: This looks good, let's merge it.
          type: string
        created_at:
          example: "2026-01-15T10:30:00Z"
          type: string
        id:
          example: c3d4e5f6-a7b8-9012-cdef-123456789012
          type: string
        item_id:
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          type: string
        updated_at:
          example: "2026-01-15T10:30:00Z"
          type: string
      type: object
    api.SwaggerCreateCommentRequest:
      properties:
        content:
          example: This looks good, let's merge it.
          type: string
      type: object
...
    api.SwaggerErrorDetail:
      properties:
        code:
          example: UNAUTHORIZED
          type: string
        message:
          example: invalid email or password
          type: string
        request_id:
          example: req_a8d9912a
          type: string
      type: object
    api.SwaggerErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/api.SwaggerErrorDetail'
      type: object
...
    notifications.listNotificationsResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/notifications.notificationResponse'
          type: array
          uniqueItems: false
        unread_count:
          type: integer
      type: object
    notifications.notificationResponse:
      properties:
        body:
          type: string
        created_at:
          type: string
        entity_id:
          type: string
        entity_kind:
          type: string
        id:
          type: string
        is_read:
          type: boolean
        kind:
          type: string
        title:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer <token>". Obtain via POST /api/v1/auth/login'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: hello@azimuthalhq.com
    name: Azimuthal HQ
    url: https://azimuthalhq.com
  description: |-
    The fully open-source alternative to Jira, Confluence, and Jira Service Desk.
    Self-hostable, single binary, Apache 2.0 licensed.

    ## Authentication
    All endpoints except /auth/login and /auth/register require a Bearer JWT token.
    Obtain a token via POST /api/v1/auth/login, then include it as:
    `Authorization: Bearer <your-token>`
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Azimuthal API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /notifications:
    get:
      description: Returns the authenticated user's notifications, unread first. Paginated
        via limit/offset query params.
      parameters:
      - description: Max results (default 20)
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notifications.listNotificationsResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{notificationID}/read:
    post:
      description: Marks the specified notification as read. Only the owning user
        can mark their own notifications.
      parameters:
      - description: Notification ID
        in: path
        name: notificationID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Mark notification read
      tags:
      - notifications
  /notifications/read-all:
    post:
      description: Marks every unread notification for the authenticated user as read.
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Mark all notifications read
      tags:
      - notifications
...
  /orgs/{orgID}/spaces/{spaceID}/items/{itemID}/comments:
    get:
      description: Returns all comments for the specified item.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.SwaggerCommentResponse'
                type: array
          description: List of comments
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid item ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: List comments
      tags:
      - comments
    post:
      description: Adds a new comment to the specified item. Author is set from the
        JWT.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerCreateCommentRequest'
                description: Comment content
                summary: body
        description: Comment content
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerCommentResponse'
          description: Comment created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - comments
...
servers:
- url: localhost:8080/api/v1